requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1 ; python_version >= '3.7'
anyio==4.0.0 ; python_version >= '3.8'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
black==23.7.0
blessed==1.20.0 ; python_version >= '2.7'
cachetools==5.3.1 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
cloudevents==1.9.0
colorful==0.5.5
cuda-python==12.2.0
cython==3.0.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
deprecation==2.1.0
diskcache==5.6.3 ; python_version >= '3'
distlib==0.3.7
fastapi==0.95.2 ; python_version >= '3.7'
filelock==3.12.3 ; python_version >= '3.8'
frozenlist==1.4.0 ; python_version >= '3.8'
google-api-core==2.11.1 ; python_version >= '3.6'
google-auth==2.22.0 ; python_version >= '3.6'
googleapis-common-protos==1.60.0 ; python_version >= '3.7'
gpustat==1.1.1
grpcio==1.51.3 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httptools==0.6.0
httpx==0.23.3 ; python_version >= '3.7'
idna==3.4
jsonschema==4.19.0 ; python_version >= '3.8'
jsonschema-specifications==2023.7.1 ; python_version >= '3.8'
kserve==0.11.0
kubernetes==27.2.0 ; python_version >= '3.6'
llama-cpp-python==0.1.83
msgpack==1.0.5
multidict==6.0.4 ; python_version >= '3.7'
mypy-extensions==1.0.0 ; python_version >= '3.5'
numpy==1.25.2 ; python_version >= '3.9'
nvidia-ml-py==12.535.108
oauthlib==3.2.2 ; python_version >= '3.6'
opencensus==0.11.2
opencensus-context==0.1.3
orjson==3.9.5 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pandas==2.1.0 ; python_version >= '3.9'
pathspec==0.11.2 ; python_version >= '3.7'
platformdirs==3.10.0 ; python_version >= '3.7'
prometheus-client==0.13.1 ; python_version >= '3.6'
protobuf==3.20.3 ; python_version >= '3.7'
psutil==5.9.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py-spy==0.3.14
pyasn1==0.5.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==1.10.12 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0 ; python_version >= '3.8'
python-rapidjson==1.10 ; python_version >= '3.6'
pytz==2023.3
pyyaml==6.0.1 ; python_version >= '3.6'
ray[serve]==2.4.0
referencing==0.30.2 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3986[idna2008]==1.5.0
rpds-py==0.10.0 ; python_version >= '3.8'
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==6.3.0
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
tabulate==0.9.0 ; python_version >= '3.7'
timing-asgi==0.3.1 ; python_version >= '3.6' and python_version < '4.0'
tritonclient==2.37.0.9383150
typing-extensions==4.7.1 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.19.0 ; python_version >= '3.7'
uvloop==0.17.0
virtualenv==20.21.0 ; python_version >= '3.7'
watchfiles==0.20.0
wcwidth==0.2.6
websocket-client==1.6.2 ; python_version >= '3.8'
websockets==11.0.3
yarl==1.9.2 ; python_version >= '3.7'
